Fix signed comparison warnings.

+++ compiler/array-helpers.cpp
@@ -525,7 +525,7 @@ FixedArrayValidator::CheckArgument(Expr* init)
     if (type_.ident == iREFARRAY)
         return true;
 
-    for (int i = 0; i < type_.dim.size(); i++) {
+    for (size_t i = 0; i < type_.dim.size(); i++) {
         if (type_.dim[i] && type_.dim[i] != dim[i]) {
             error(expr->pos(), 48);
             return false;
